+++ PKHeX.Core.AutoMod/AutoMod/Regeneration/RegenTemplateExtensions.cs
@@ -5,7 +5,7 @@ namespace PKHeX.Core.AutoMod
 {
     public static class RegenTemplateExtensions
     {
-        public static void SanitizeForm(this RegenTemplate set, int gen)
+        public static void SanitizeForm(this RegenTemplate set, byte gen)
         {
                 // Scatterbug and Spewpa must be Fancy - not anymore
             if (!FormInfo.IsBattleOnlyForm(set.Species, set.Form, gen))
